# mysql/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto Integrador",
        "version": "1.0.0",
        "description": "Documentação do Projeto Integrador"
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
    "paths": {
        "/produtos": {
            "get": {
                "summary": "Retorna todos os produtos",
                "tags": [
                    "Produtos"
                ],
                "responses": {
                    "200": {
                        "description": "Uma lista de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "idProduct": {
                                                "type": "integer",
                                                "description": "ID do produto"
                                            },
                                            "nameProduct": {
                                                "type": "string",
                                                "description": "Nome do produto"
                                            },
                                            "sizeProduct": {
                                                "type": "string",
                                                "description": "Tamanhos do produto"
                                            },
                                            "priceProduct": {
                                                "type": "number",
                                                "description": "Preço do produto"
                                            },
                                            "imageProduct": {
                                                "type": "string",
                                                "description": "Caminho da imagem do produto"
                                            },
                                            "descriptionProduct": {
                                                "type": "string",
                                                "description": "Descrição do produto"
                                            },
                                            "categoryProduct": {
                                                "type": "string",
                                                "description": "Categoria do produto"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar produtos"
                    }
                }
            }
        },
        "/produto/{id}": {
            "get": {
                "summary": "Retorna os detalhes de um produto específico",
                "description": "Busca os detalhes de um produto baseado no ID fornecido.",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do produto",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "idProduct": {
                                            "type": "integer",
                                            "description": "ID do produto"
                                        },
                                        "nameProduct": {
                                            "type": "string",
                                            "description": "Nome do produto"
                                        },
                                        "sizeProduct": {
                                            "type": "string",
                                            "description": "Tamanhos do produto"
                                        },
                                        "priceProduct": {
                                            "type": "number",
                                            "description": "Preço do produto"
                                        },
                                        "imageProduct": {
                                            "type": "string",
                                            "description": "Caminho da imagem do produto"
                                        },
                                        "descriptionProduct": {
                                            "type": "string",
                                            "description": "Descrição do produto"
                                        },
                                        "categoryProduct": {
                                            "type": "string",
                                            "description": "Categoria do produto"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Produto não encontrado"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar detalhes do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Erro ao buscar detalhes do produto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/produtos": {
            "get": {
                "summary": "Lista produtos por categoria",
                "description": "Retorna uma lista de produtos filtrados por categoria.",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": true,
                        "description": "Categoria dos produtos a serem buscados.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "idProduct": {
                                                "type": "integer",
                                                "description": "ID do produto"
                                            },
                                            "nameProduct": {
                                                "type": "string",
                                                "description": "Nome do produto"
                                            },
                                            "sizeProduct": {
                                                "type": "string",
                                                "description": "Tamanhos do produto"
                                            },
                                            "priceProduct": {
                                                "type": "number",
                                                "description": "Preço do produto"
                                            },
                                            "imageProduct": {
                                                "type": "string",
                                                "description": "Caminho da imagem do produto"
                                            },
                                            "descriptionProduct": {
                                                "type": "string",
                                                "description": "Descrição do produto"
                                            },
                                            "categoryProduct": {
                                                "type": "string",
                                                "description": "Categoria do produto"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar produtos."
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "summary": "Busca produtos por nome",
                "description": "Retorna uma lista de produtos cujo nome contém o termo de busca fornecido.",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "busca",
                        "in": "busca",
                        "required": true,
                        "description": "Termo a ser buscado no nome dos produtos.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos correspondente à busca.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "idProduct": {
                                                "type": "integer",
                                                "description": "ID do produto"
                                            },
                                            "nameProduct": {
                                                "type": "string",
                                                "description": "Nome do produto"
                                            },
                                            "sizeProduct": {
                                                "type": "string",
                                                "description": "Tamanhos do produto"
                                            },
                                            "priceProduct": {
                                                "type": "number",
                                                "description": "Preço do produto"
                                            },
                                            "imageProduct": {
                                                "type": "string",
                                                "description": "Caminho da imagem do produto"
                                            },
                                            "descriptionProduct": {
                                                "type": "string",
                                                "description": "Descrição do produto"
                                            },
                                            "categoryProduct": {
                                                "type": "string",
                                                "description": "Categoria do produto"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Necessita de um termo de busca."
                    },
                    "500": {
                        "description": "Erro ao realizar a busca no banco de dados."
                    }
                }
            }
        },
        "/api/cadastroUser": {
            "post": {
                "summary": "Cadastro de usuário",
                "description": "Cria um novo usuário no sistema com email, nome e senha.",
                "tags": [
                    "Autenticação"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "description": "Nome completo do usuário",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "Email do usuário",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "Senha do usuário",
                                        "example": "senha123"
                                    }
                                },
                                "required": [
                                    "nome",
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Usuário cadastrado com sucesso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos fornecidos."
                    },
                    "500": {
                        "description": "Erro interno do servidor ao processar a solicitação."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login de usuário",
                "description": "Realiza o login do usuário e retorna um token JWT para autenticação.",
                "tags": [
                    "Autenticação"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "description": "Email do usuário",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "Senha do usuário",
                                        "example": "senha123"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso. Retorna um token JWT e detalhes do usuário.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "description": "Token JWT para autenticação."
                                        },
                                        "username": {
                                            "type": "string",
                                            "description": "Nome do usuário logado."
                                        },
                                        "isAdmin": {
                                            "type": "boolean",
                                            "description": "Indica se o usuário é um administrador."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não encontrado ou senha incorreta."
                    },
                    "500": {
                        "description": "Erro interno do servidor ao processar a solicitação."
                    }
                }
            }
        },
...
    }
}